refactor: Replace mangoIsPublishedOrScheduled with mangoIsPublished to streamline publish date handling and include scheduled content option
