+++ app/src/components/HomePage/ContinueWatching.vue
@@ -3,7 +3,7 @@
 import HorizontalContentTileCollection from "@/components/content/HorizontalContentTileCollection.vue";
 import { type ContentDto, DocType, db, PostType, TagType, type Uuid, mangoToDexie, useDexieLiveQueryWithDeps } from "luminary-shared";
 import { appLanguageIdsAsRef } from "@/globalConfig";
-import { mangoIsPublishedOrScheduled } from "@/util/mangoIsPublished";
+import { mangoIsPublished } from "@/util/mangoIsPublished";
 import { useI18n } from "vue-i18n";
 import { ref, onMounted, onUnmounted } from "vue";
 
@@ -67,7 +67,7 @@ const watchedContent = useDexieLiveQueryWithDeps(
                     { type: DocType.Content },
                     { parentPostType: { $ne: PostType.Page } },
                     { parentTagType: { $ne: TagType.Category } },
-                    ...mangoIsPublishedOrScheduled(appLanguageIds),
+                    ...mangoIsPublished(appLanguageIds, { includeScheduled: true }),
                 ],
             },
         });
